:art: Split out streaming step, about to refactor
